Break out middleware to allow testing without dependency on AWS
