+++ app/server/apiProxy.ts
@@ -7,10 +7,8 @@ import { X_GU_ID_FORWARDED_SCOPE } from '../shared/identity';
 import { MDA_TEST_USER_HEADER } from '../shared/productResponse';
 import { conf } from './config';
 import { log, putMetric } from './log';
-import {
-	augmentRedirectURL,
-	getCookiesOrEmptyString,
-} from './middleware/identityMiddleware';
+import { getCookiesOrEmptyString } from './middleware/identityMiddleware';
+import { augmentRedirectURL } from './middleware/requestMiddleware';
 
 type BodyHandler = (res: Response, body: Buffer) => void;
 type JsonString = Buffer | string | undefined;
